updating examples
